LayoutTests/fast/css/parsing-a98rgb-colors.html

<html>
    <script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
    description("Test the parsing of lab, lch and gray colors.");

    function computedStyle(property, value)
    {
        var div = document.createElement("div");
        document.body.appendChild(div);
        div.style.setProperty(property, value);
        var computedValue = getComputedStyle(div).getPropertyValue(property);
        document.body.removeChild(div);
        return computedValue;
    }

    function innerStyle(property, value)
    {
        var div = document.createElement("div");
        div.style.setProperty(property, value);
        return div.style.getPropertyValue(property);
    }

    function testComputed(property, value, expected)
    {
        shouldBeEqualToString('computedStyle("' + property + '", "' + value + '")', expected);
    }

    function testInner(property, value, expected)
    {
        if (expected === null)
            expected = "";
        shouldBeEqualToString('innerStyle("' + property + '", "' + value + '")', expected);
    }

    testComputed("background-color", "color(a98-rgb 0% 0% 0%)", "color(a98-rgb 0 0 0)");
    testComputed("background-color", "color(a98-rgb 10% 10% 10%)", "color(a98-rgb 0.1 0.1 0.1)");
    testComputed("background-color", "color(a98-rgb .2 .2 25%)", "color(a98-rgb 0.2 0.2 0.25)");
    testComputed("background-color", "color(a98-rgb 0 0 0 / 1)", "color(a98-rgb 0 0 0)");
    testComputed("background-color", "color(a98-rgb 0% 0 0 / 0.5)", "color(a98-rgb 0 0 0 / 0.5)");
    testComputed("background-color", "color(a98-rgb 20% 0 10/0.5)", "color(a98-rgb 0.2 0 1 / 0.5)");
    testComputed("background-color", "color(a98-rgb 20% 0 10/50%)", "color(a98-rgb 0.2 0 1 / 0.5)");
    testComputed("background-color", "color(a98-rgb 400% 0 10/50%)", "color(a98-rgb 1 0 1 / 0.5)");
    testComputed("background-color", "color(a98-rgb 50% -160 160)", "color(a98-rgb 0.5 0 1)");
    testComputed("background-color", "color(a98-rgb 50% -200 200)", "color(a98-rgb 0.5 0 1)");
    testComputed("background-color", "color(a98-rgb 0 0 0 / -10%)", "color(a98-rgb 0 0 0 / 0)");
    testComputed("background-color", "color(a98-rgb 0 0 0 / 110%)", "color(a98-rgb 0 0 0)");
    testComputed("background-color", "color(a98-rgb 0 0 0 / 300%)", "color(a98-rgb 0 0 0)");
    testComputed("background-color", "color(a98-rgb 50% -200)", "color(a98-rgb 0.5 0 0)");
    testComputed("background-color", "color(a98-rgb 50%)", "color(a98-rgb 0.5 0 0)");
    testComputed("background-color", "color(a98-rgb)", "color(a98-rgb 0 0 0)");
    testComputed("background-color", "color(a98-rgb 50% -200 / 0.5)", "color(a98-rgb 0.5 0 0 / 0.5)");
    testComputed("background-color", "color(a98-rgb 50% / 0.5)", "color(a98-rgb 0.5 0 0 / 0.5)");
    testComputed("background-color", "color(a98-rgb / 0.5)", "color(a98-rgb 0 0 0 / 0.5)");

    debug('');
    debug('Test invalid values');
    testComputed("background-color", "color(a98-rgb 0 0 0 0)", "rgba(0, 0, 0, 0)");
    testComputed("background-color", "color(a98-rgb 0deg 0% 0)", "rgba(0, 0, 0, 0)");
    testComputed("background-color", "color(a98-rgb 0% 0 0 1)", "rgba(0, 0, 0, 0)");
    testComputed("background-color", "color(a98-rgb 0% 0 0 10%)", "rgba(0, 0, 0, 0)");
    testComputed("background-color", "color(a98-rgb 0% 0 0deg)", "rgba(0, 0, 0, 0)");
    testComputed("background-color", "color(a98-rgb 0% 0% 0deg)", "rgba(0, 0, 0, 0)");
    testComputed("background-color", "color(a98-rgb 40% 0 0deg)", "rgba(0, 0, 0, 0)");

</script>
    
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
