LayoutTests/fast/css/parsing-a98rgb-colors-expected.txt

Test the parsing of lab, lch and gray colors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle("background-color", "color(a98-rgb 0% 0% 0%)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 10% 10% 10%)") is "color(a98-rgb 0.1 0.1 0.1)"
PASS computedStyle("background-color", "color(a98-rgb .2 .2 25%)") is "color(a98-rgb 0.2 0.2 0.25)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / 1)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0 / 0.5)") is "color(a98-rgb 0 0 0 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 20% 0 10/0.5)") is "color(a98-rgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 20% 0 10/50%)") is "color(a98-rgb 0.2 0 1 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 400% 0 10/50%)") is "color(a98-rgb 1 0 1 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 50% -160 160)") is "color(a98-rgb 0.5 0 1)"
PASS computedStyle("background-color", "color(a98-rgb 50% -200 200)") is "color(a98-rgb 0.5 0 1)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / -10%)") is "color(a98-rgb 0 0 0 / 0)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / 110%)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 / 300%)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 50% -200)") is "color(a98-rgb 0.5 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 50%)") is "color(a98-rgb 0.5 0 0)"
PASS computedStyle("background-color", "color(a98-rgb)") is "color(a98-rgb 0 0 0)"
PASS computedStyle("background-color", "color(a98-rgb 50% -200 / 0.5)") is "color(a98-rgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb 50% / 0.5)") is "color(a98-rgb 0.5 0 0 / 0.5)"
PASS computedStyle("background-color", "color(a98-rgb / 0.5)") is "color(a98-rgb 0 0 0 / 0.5)"

Test invalid values
PASS computedStyle("background-color", "color(a98-rgb 0 0 0 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0deg 0% 0)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0 1)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0 10%)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 0% 0% 0deg)") is "rgba(0, 0, 0, 0)"
PASS computedStyle("background-color", "color(a98-rgb 40% 0 0deg)") is "rgba(0, 0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
